pyproject.toml
[project]
name = "front-end"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "duckdb>=1.0.0",
    "pandas>=2.2.2",
    "pyarrow>=22.0.0",
    "python-dotenv>=1.2.1",
    "streamlit>=1.37.1",
]
